+++ cli/server.go
@@ -62,9 +62,9 @@ import (
 	"github.com/coder/wgtunnel/tunnelsdk"
 
 	"github.com/coder/coder/v2/coderd/entitlements"
-	"github.com/coder/coder/v2/coderd/notifications/push"
 	"github.com/coder/coder/v2/coderd/notifications/reports"
 	"github.com/coder/coder/v2/coderd/runtimeconfig"
+	"github.com/coder/coder/v2/coderd/webpush"
 
 	"github.com/coder/coder/v2/buildinfo"
 	"github.com/coder/coder/v2/cli/clilog"
@@ -779,17 +779,17 @@ func (r *RootCmd) Server(newAPI func(context.Context, *coderd.Options) (*coderd.
 			// Manage push notifications.
 			experiments := coderd.ReadExperiments(options.Logger, options.DeploymentValues.Experiments.Value())
 			if experiments.Enabled(codersdk.ExperimentWebPush) {
-				webpusher, err := push.New(ctx, &options.Logger, options.Database)
+				webpusher, err := webpush.New(ctx, &options.Logger, options.Database)
 				if err != nil {
 					options.Logger.Error(ctx, "failed to create web push dispatcher", slog.Error(err))
 					options.Logger.Warn(ctx, "web push notifications will not work until the VAPID keys are regenerated")
-					webpusher = &push.NoopWebpusher{
+					webpusher = &webpush.NoopWebpusher{
 						Msg: "Web Push notifications are disabled due to a system error. Please contact your Coder administrator.",
 					}
 				}
 				options.WebpushDispatcher = webpusher
 			} else {
-				options.WebpushDispatcher = &push.NoopWebpusher{
+				options.WebpushDispatcher = &webpush.NoopWebpusher{
 					Msg: "Web Push notifications are an experimental feature and are disabled by default. Enable the 'web-push' experiment to use this feature.",
 				}
 			}
